remove startup file & move to test
