+++ data/tests/views_test.py
@@ -1,5 +1,29 @@
 import pytest
 import requests
+import subprocess
+# from os import system
+# import platform
+
+def startServer():
+    makeMigrationsProc = subprocess.Popen(["python", "manage.py", "makemigrations", "data"], shell=True)
+    makeMigrationsProc.wait()
+    migrateProc = subprocess.Popen(["python", "manage.py", "migrate", "--database=testing"], shell=True)
+    migrateProc.wait()
+    runserverProcess = subprocess.Popen(["python", "manage.py" ,"runserver", "--settings=BackendDev.testSettings"], shell=True)
+    return(runserverProcess)
+
+def stopServer(runServerProcess):
+    runServerProcess.terminate()
+    # try:
+    #     if(platform.system() == "Windows"):
+    #             system("del databases/db.sqlite3")
+    #     else:
+    #             system("rm databases/db.sqlite3")
+    # except Exception as e:
+    #     print("Exception: " + e)
+
+# proc = startServer()
+# stopServer(proc)
 
 # NOTE: make sure the server is running before running tests!
 # these tests assume that there is nothing in the database beforehand and you are running them in order
